requirements.txt
amqp==1.4.6
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.19
celery==3.1.17
certifi==14.05.14
kombu==3.0.24
pika==0.9.14
pytz==2014.10
tornado==4.1
tornado-celery==0.3.4
upyun==2.2.5
wsgiref==0.1.2
